refactor(modal): update ModalForgotPassword and ModalSuccess to use new InputField and PrimaryButton
